Install script ready for WSL

+++ install.sh
@@ -1,4 +1,4 @@
-e#!/bin/bash -e
+#!/bin/bash -e
 
 # Install Script based on Optixal's neovim-init.vim install script
 # Use on Ubuntu 22.04 LTS
@@ -25,11 +25,31 @@ mkdir -p ~/.local/bin
 tar xf /tmp/nvim-linux64.tar.gz -C ~/.local
 ln -sf $(readlink -f ~/.local/nvim-linux64/bin/nvim) ~/.local/bin/nvim
 
+# Add ~/.local/bin to PATH if it's not already in it
+if ! [[ ":$PATH:" == *":$HOME/.local/bin:"* ]]; then
+    echo "[*] Adding ~/.local/bin to PATH"
+    if [ -n "`$SHELL -c 'echo $ZSH_VERSION'`" ]; then
+        SHELL_CONFIG_FILE=~/.zshrc
+    elif [ -n "`$SHELL -c 'echo $BASH_VERSION'`" ]; then
+        SHELL_CONFIG_FILE=~/.profile
+    else
+        echo "[-] Could not detect what shell you are using. Ensure to manually add ~/.local/bin to your PATH"
+    fi
+    echo -e '\nPATH="$HOME/.local/bin:$PATH"' >> $SHELL_CONFIG_FILE
+    export PATH="$HOME/.local/bin:$PATH"
+fi
+
 sudo apt install \
 	nodejs \
 	fd-find \
+	ripgrep \
+	cmake \
 	-y	
-ln -s $(which fdfind) ~/.local/bin/fd
+if [ -e ~/.local/bin/fd ]; then
+    echo "fd exists."
+else
+    ln -s $(which fdfind) ~/.local/bin/fd
+fi
 
 # Install vim-plug plugin manager
 echo '[*] Installing vim-plug'
@@ -37,9 +57,9 @@ sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.
 
 # Enter Neovim and install plugins with vim-plug's :PlugInstall using a temporary init.vim, which avoids warnings about missing colorschemes, functions, etc
 echo -e '[*] Running :PlugInstall within nvim ...'
-sed '/call plug#end/q' ../init.vim > ~/.config/nvim/init.vim
+sed '/call plug#end/q' ./init.vim > ~/.config/nvim/init.vim
 nvim -c 'PlugInstall' -c 'qa'
 
 # Copy init.vim and lua scripts in current working directory to nvim's config location
 echo '[*] Copying init.vim & lua/ -> ~/.config/nvim/'
-cp -r ../init.vim ../lua/ ~/.config/nvim/
+cp -r ./init.vim ./lua/ ~/.config/nvim/
